refactor(observability): route OTel bootstrap/lifecycle through setup
Point the app, Celery app, task bootstrap, and perf timer at the new
setup/lifecycle + signals entrypoints instead of the old bootstrap module.
